Add possibility to distinguish create and edit options
Add predicat function to determine whether we are creating or updating
object and special field in description of entity fields describing when
should apply this option. It is needed, for example, to edit password
field. When new user gets created, password is required. In editing case,
password can be blank.