Add composer.lock file
It's necessary because this file locks the project to those specific versions of dependencies.
This means that anyone who sets up the project will download the exact same version of the dependencies.
Everyone runs on the same dependencies, which mitigates the potential for bugs affecting only some parts of the deployments.